Moved type definitions to their own file
