+++ packages/opencensus-exporter-azure/src/azure-stats.ts
@@ -15,102 +15,26 @@
  */
 
 import {
-    ExporterConfig,
     StatsEventListener,
     View,
     Measurement,
     TagKey,
     TagValue,
     logger,
-    Logger,
-    Metric,
-    MetricProducerManager,
-    Metrics,
-    Measure,
-    AggregationData,
-    MetricDescriptorType,
 } from '@opencensus/core';
+
+import {
+    AzureStatsExporterOptions,
+    ExportMode,
+    AggregationMethod,
+    IllegalOptionsError
+} from './types';
 
 import * as ApplicationInsights from 'applicationinsights';
 
 import {
     MetricTelemetry
 } from 'applicationinsights/out/Declarations/Contracts';
-
-export interface StatsParams {
-    registeredViews: View[];
-    registeredMeasures: Measure[];
-    recordedData: { [key: string]: AggregationData[] };
-}
-
-export enum ExportMode {
-    SINGLE_VALUE = 0,
-    BATCH = 1
-}
-
-export enum AggregationMethod {
-    AVERAGE = 0,
-    MIN = 1,
-    MAX = 2,
-    SUM = 3,
-    COUNT = 4
-}
-
-/**
- * Options for Azure Monitor configuration.
- */
-export interface AzureStatsExporterOptions extends ExporterConfig {
-
-    /**
-     * The Instrumentation Key found in your application's Azure Monitor Application Insights
-     * Overview page. Required.
-     */
-    instrumentationKey: string;
-
-    /**
-     * If specified, dictates the mode the exporter will function in.
-     * Optional, defaults to ExportMode.SINGLE_VALUE.
-     */
-    exportMode?: ExportMode;
-
-    /**
-     * If specified, defines the number of milliseconds between uploading metrics
-     * to Azure Monitor. Optional, defaults to 60,000 (1 minute).
-     */
-    periodInMillis?: number;
-
-    /**
-     * If specified, this will override the default OpenCensus prefix of an
-     * Azure Monitor metric. Optional.
-     */
-    prefix?: string;
-
-    /**
-     * The aggregation method if batch exporting is to be used.
-     * Optional, defaults to AggregationMethod.AVERAGE.
-     */
-    aggregationMethod?: AggregationMethod;
-
-    /**
-     * If specified, this will serve as the logger used by the exporter.
-     * Optional, default to use whatever logger is registered with OpenCensus.
-     */
-    logger?: Logger;
-
-    /**
-     * If specified, this function will be called whenever an error occurs uploading
-     * stats to Azure monitor. Optional.
-     */
-    onMetricUploadError?: (err: Error) => void;
-
-}
-
-export class IllegalOptionsError extends Error {
-    constructor(message: string) {
-        super(message);
-        this.name = 'IllegalOptionsError';
-    }
-}
 
 /**
  * Configuration defaults for an AzureStatsExporter.
